ajustes na mana que ocasionaram varios bags visuais

+++ src/main/java/com/github/uranus_mod_group/uranus_mod/client/ClientManaData.java
@@ -1,15 +1,15 @@
 package com.github.uranus_mod_group.uranus_mod.client;
 
 public class ClientManaData {
-    private static int playerMana;
+    private static double playerMana;
 
     //set mana on playerMana
-    public static void set(int mana)
+    public static void set(double mana)
     {
         ClientManaData.playerMana = mana;
     }
 
-    public static int getPlayerMana()
+    public static double getPlayerMana()
     {
         return playerMana;
     }
